// src/api/generated/schemas/authResponse.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * NervesCloud API
 * The NervesCloud API gives users full access to their
Orgs, Products, and corresponding Device fleets.

The API can be used to integrate with your own systems, providing full access to your Product and Device data.

The API is documented using the OpenAPI 3.0 specification.

 * OpenAPI spec version: 2.0.0
 */
import type { AuthResponseData } from './authResponseData';

export interface AuthResponse {
  data?: AuthResponseData;
}